Propagate 'at' through federated resolvers and certify it
The root query's 'at' timestamp was dropped at every resolver hop —
all six resolver paths (HTTP and registry, singleton and vector,
internal variants) evaluated at Utc::now(), so point-in-time reads
returned current-state related entities. The sibling meshql-ts and
meshql-java suites certify this propagation; the Rust port's
farm.feature even carried the orphaned setup for the missing scenario
(capture first_stamp, rename coop red->purple) without the scenario
itself.

Fix: stamp the effective 'at' onto every resolved parent object under
an internal __at key (the __ prefix is reserved by GraphQL so it can
never collide with a schema field) and have resolvers read it instead
of now, re-stamping their results so deep chains stay consistent.

Certification:
- farm.feature: restore the temporal federated scenario (coops at
  first_stamp are red, not purple) plus a latest-state control
- egg_economy.feature: one-hop (internal resolver) and two-hop deep
  federation temporal scenarios
- new contains/not-contains step defs with array-flattening path
  navigation for asserting on nested federated collections

Also fix 'at: Int' -> 'at: Float' across cert fixtures and farm
example schemas: GraphQL Int is 32-bit and millisecond timestamps
overflow it (egg-economy and meshql-mcp already used Float). Skill
docs updated to match the restored guarantee.

+++ meshql-postgres/tests/farm_cert.rs
@@ -21,8 +21,8 @@ type Coop {
     name: String
 }
 type Query {
-    getFarm(id: ID, at: Int): Farm
-    getFarms(name: String, at: Int): [Farm]
+    getFarm(id: ID, at: Float): Farm
+    getFarms(name: String, at: Float): [Farm]
 }
 "#;
 
@@ -37,9 +37,9 @@ type Coop {
 type Farm { id: ID name: String }
 type Hen { id: ID name: String }
 type Query {
-    getCoop(id: ID, at: Int): Coop
-    getCoops(name: String, at: Int): [Coop]
-    getCoopsByFarm(id: ID, at: Int): [Coop]
+    getCoop(id: ID, at: Float): Coop
+    getCoops(name: String, at: Float): [Coop]
+    getCoopsByFarm(id: ID, at: Float): [Coop]
 }
 "#;
 
@@ -53,9 +53,9 @@ type Hen {
 }
 type Coop { id: ID name: String }
 type Query {
-    getHen(id: ID, at: Int): Hen
-    getHens(name: String, at: Int): [Hen]
-    getHensByCoop(id: ID, at: Int): [Hen]
+    getHen(id: ID, at: Float): Hen
+    getHens(name: String, at: Float): [Hen]
+    getHensByCoop(id: ID, at: Float): [Hen]
 }
 "#;
